This is just my base starting point. All you can do is (slowly) move the paddle with the ABXY buttons, and change it's size with the D-Pad (was used for testing good paddle sizes.) And unlike before (and unlike my other projects that say it) there actually is a red area on the bottom screen AND you actually can touch it to close the application.
Obviously this isn't even remotely in the playable stage. But you can have a little fun messing around with it I suppose. More so if you are also a 3ds homebrew developer and you're just starting out with (2d) graphics like me, in which case, this is the most barebones it could be.

(Just remember than you need sf2dlib 'installed' and possibly citro3d.)