Bump for Beta 1
Breakout is now open for Beta testing (techinically it always was, and probably is, but now is actually the time I'm considering to be the testing period.)

And yes I know you can't hit the bricks okay. I haven't added that yet.

Appearantly this is my list of priorities:
*Working application
*paddle hit physics to change direction (and soon speed)
*bricks

I think 2 and 3 should probably be swaped... RIP.