Made faster
It was lagging, with the new thing I added for checking bricks, so I made the for loop shorter. I also optimized the code slightly, and I have noticed that the ball to paddle detection is still a little weird in some places, but I don't know why, nor do I know how to fix it at this time.

Also don't question the recent commit chain if it looks wacky. I have no idea what happened. I made the readme file online on github.com then started working on the project without pulling the commit containing the readme.md to my computer, so when i went to push the new commit, I had an error, and a bunch of shit happened, and I had to reset my local files to the readme commit, (after backing them up) then I restored them to make this commit. So yeah... the fact that I couldn't pull a single commit with one added file, is bullshit, github please fix that, I have no idea what just happened and it was very frustrating and time consuming. Time = Money!