Laser paddle design (may change)
One funny story I'll put here to remember later down the line in life:

While developing the laser powerup I got to a point where when I was first testing it, it would shoot and destroy the block , but it would never hit any thing past the first row, because it would never get past the first row and it took me a surprising amount of time to realize I hadn't told it to check if the brick existed or not before destroying the laser. Pretty big facepalm moment.