Unstable Commit
I'm going to change this soon.
The ball's hit detection is now a rectangle (square) but still glitches. Every frame (not every detection, but every frame) a variable of direction, --scratch that--

New idea:
every time a brick is hit, change angle by 90 (I'll have to figure out + or -) but keep a vector of which bricks the ball is inside of, and if it's still inside the brick on the next frame, don't change angle again, wait until no longer in brick to remove value from vector.